Initial code commit, WIP but basic use is possible
Links are not yet fully fleshed out for use
The main idea for this is that whichever classes you need will inherit
StateNode and all data that is important to the state, will be used with
member literals.
Easier access is needed to make that less painful but
that is yet to come. (I imagine something like overloading the __getattr__
function in python)
Another important thing that is missing is more complex datastructures
like lists or whatever but they should actually not be too complicated,
just needs some string in and out, and a lot of classes... Or more
playing around with generics, which is more likely cause that's where
the fun starts ^-^
We'll see about time and motivation though